pyproject.toml
[project]
name = "adult_income"
version = "0.1.0"
description = "Initial project: Adult Income"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "hydra-core>=1.3.2",
    "ipython>=9.5.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "mlflow>=3.6.0",
    "notebook>=7.4.7",
    "numpy>=2.3.3",
    "optuna>=4.6.0",
    "pandas>=2.3.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.1",
]
